+++ qsiprep/data/intramodal_ACPC.json
@@ -1,0 +1,25 @@
+{
+    "dimension": 3,
+    "float": true,
+    "winsorize_lower_quantile": 0.002,
+    "winsorize_upper_quantile": 0.998,
+    "collapse_output_transforms": true,
+    "write_composite_transform": false,
+    "initial_moving_transform_com": 1,
+    "use_histogram_matching": [true, true],
+    "transforms": [ "Similarity", "Affine"],
+    "number_of_iterations": [ [ 100000, 100000 ], [ 10000, 10000 ] ],
+    "output_warped_image": true,
+    "transform_parameters": [ [ 0.3 ], [ 0.1 ] ],
+    "convergence_threshold": [ 1e-6, 1e-6 ],
+    "convergence_window_size": [ 10, 10 ],
+    "metric": [ "MI", "MI" ],
+    "sampling_percentage": [ 0.7, 0.2 ],
+    "sampling_strategy": [ "Regular", "Regular" ],
+    "shrink_factors": [ [ 4, 2 ], [ 2, 1 ] ],
+    "sigma_units": [ "vox", "vox" ],
+    "metric_weight": [ 1, 1 ],
+    "smoothing_sigmas": [ [ 0, 0 ], [ 2, 2 ] ],
+    "radius_or_number_of_bins": [ 32, 32],
+    "interpolation": "Linear"
+}
