+++ qsiprep/workflows/anatomical/__init__.py
@@ -1,0 +1,3 @@
+from .volume import (init_synthstrip_wf, init_anat_preproc_wf,
+                     TEMPLATE_MAP, init_dwi_reference_wf, init_dl_prep_wf,
+                     init_synthseg_wf)
