+++ qsiprep/workflows/anatomical/surface.py
@@ -1,0 +1,693 @@
+# -*- coding: utf-8 -*-
+# emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
+# vi: set ft=python sts=4 ts=4 sw=4 et:
+"""
+Anatomical reference preprocessing workflows
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. autofunction:: init_anat_preproc_wf
+.. autofunction:: init_skullstrip_ants_wf
+
+"""
+from pkg_resources import resource_filename as pkgr
+
+from nipype.pipeline import engine as pe
+from nipype.interfaces import (
+    io as nio,
+    utility as niu,
+    c3,
+    freesurfer as fs,
+    fsl,
+    image,
+    ants,
+    afni
+)
+from nipype.interfaces.ants import BrainExtraction, N4BiasFieldCorrection
+
+from ...niworkflows.interfaces.registration import RobustMNINormalizationRPT
+from ...niworkflows.interfaces.masks import ROIsPlot
+from ...niworkflows.interfaces.freesurfer import RobustRegister
+from ...niworkflows.interfaces.segmentation import ReconAllRPT
+from ..dwi.hmc import init_b0_hmc_wf
+from ...niworkflows.interfaces.fixes import FixHeaderApplyTransforms as ApplyTransforms
+
+from ...engine import Workflow
+from ...interfaces import (
+    StructuralReference, MakeMidthickness, FSInjectBrainExtracted,
+    FSDetectInputs, NormalizeSurf, TemplateDimensions,
+    ConcatAffines, RefineBrainMask, DerivativesDataSink as FDerivativesDataSink
+)
+
+from qsiprep.interfaces import Conform
+from ...utils.misc import fix_multi_T1w_source_name, add_suffix
+from ...interfaces.freesurfer import (
+        PatchedLTAConvert as LTAConvert, PrepareSynthStripGrid,
+        FixHeaderSynthStrip)
+from ...interfaces.anatomical import FakeSegmentation, DesaturateSkull, CustomApplyMask
+
+from nipype import logging
+LOGGER = logging.getLogger('nipype.workflow')
+
+
+class DerivativesDataSink(FDerivativesDataSink):
+    out_path_base = "qsiprep"
+
+
+TEMPLATE_MAP = {
+    'MNI152NLin2009cAsym': 'mni_icbm152_nlin_asym_09c',
+    }
+
+
+def init_surface_recon_wf(omp_nthreads, hires, name='surface_recon_wf'):
+    r"""
+    This workflow reconstructs anatomical surfaces using FreeSurfer's ``recon-all``.
+    Reconstruction is performed in three phases.
+    The first phase initializes the subject with T1w and T2w (if available)
+    structural images and performs basic reconstruction (``autorecon1``) with the
+    exception of skull-stripping.
+    For example, a subject with only one session with T1w and T2w images
+    would be processed by the following command::
+        $ recon-all -sd <output dir>/freesurfer -subjid sub-<subject_label> \
+            -i <bids-root>/sub-<subject_label>/anat/sub-<subject_label>_T1w.nii.gz \
+            -T2 <bids-root>/sub-<subject_label>/anat/sub-<subject_label>_T2w.nii.gz \
+            -autorecon1 \
+            -noskullstrip
+    The second phase imports an externally computed skull-stripping mask.
+    This workflow refines the external brainmask using the internal mask
+    implicit the the FreeSurfer's ``aseg.mgz`` segmentation,
+    to reconcile ANTs' and FreeSurfer's brain masks.
+    First, the ``aseg.mgz`` mask from FreeSurfer is refined in two
+    steps, using binary morphological operations:
+      1. With a binary closing operation the sulci are included
+         into the mask. This results in a smoother brain mask
+         that does not exclude deep, wide sulci.
+      2. Fill any holes (typically, there could be a hole next to
+         the pineal gland and the corpora quadrigemina if the great
+         cerebral brain is segmented out).
+    Second, the brain mask is grown, including pixels that have a high likelihood
+    to the GM tissue distribution:
+      3. Dilate and substract the brain mask, defining the region to search for candidate
+         pixels that likely belong to cortical GM.
+      4. Pixels found in the search region that are labeled as GM by ANTs
+         (during ``antsBrainExtraction.sh``) are directly added to the new mask.
+      5. Otherwise, estimate GM tissue parameters locally in  patches of ``ww`` size,
+         and test the likelihood of the pixel to belong in the GM distribution.
+    This procedure is inspired on mindboggle's solution to the problem:
+    https://github.com/nipy/mindboggle/blob/7f91faaa7664d820fe12ccc52ebaf21d679795e2/mindboggle/guts/segment.py#L1660
+    The final phase resumes reconstruction, using the T2w image to assist
+    in finding the pial surface, if available.
+    See :py:func:`~qsiprep.workflows.anatomical.init_autorecon_resume_wf` for details.
+    Memory annotations for FreeSurfer are based off `their documentation
+    <https://surfer.nmr.mgh.harvard.edu/fswiki/SystemRequirements>`_.
+    They specify an allocation of 4GB per subject. Here we define 5GB
+    to have a certain margin.
+    .. workflow::
+        :graph2use: orig
+        :simple_form: yes
+        from qsiprep.workflows.anatomical import init_surface_recon_wf
+        wf = init_surface_recon_wf(omp_nthreads=1, hires=True)
+    **Parameters**
+        omp_nthreads : int
+            Maximum number of threads an individual process may use
+        hires : bool
+            Enable sub-millimeter preprocessing in FreeSurfer
+    **Inputs**
+        t1w
+            List of T1-weighted structural images
+        t2w
+            List of T2-weighted structural images (only first used)
+        flair
+            List of FLAIR images
+        skullstripped_t1
+            Skull-stripped T1-weighted image (or mask of image)
+        ants_segs
+            Brain tissue segmentation from ANTS ``antsBrainExtraction.sh``
+        corrected_t1
+            INU-corrected, merged T1-weighted image
+        subjects_dir
+            FreeSurfer SUBJECTS_DIR
+        subject_id
+            FreeSurfer subject ID
+    **Outputs**
+        subjects_dir
+            FreeSurfer SUBJECTS_DIR
+        subject_id
+            FreeSurfer subject ID
+        t1_2_fsnative_forward_transform
+            LTA-style affine matrix translating from T1w to FreeSurfer-conformed subject space
+        t1_2_fsnative_reverse_transform
+            LTA-style affine matrix translating from FreeSurfer-conformed subject space to T1w
+        surfaces
+            GIFTI surfaces for gray/white matter boundary, pial surface,
+            midthickness (or graymid) surface, and inflated surfaces
+        out_brainmask
+            Refined brainmask, derived from FreeSurfer's ``aseg`` volume
+        out_aseg
+            FreeSurfer's aseg segmentation, in native T1w space
+        out_aparc
+            FreeSurfer's aparc+aseg segmentation, in native T1w space
+        out_report
+            Reportlet visualizing quality of surface alignment
+    **Subworkflows**
+        * :py:func:`~qsiprep.workflows.anatomical.init_autorecon_resume_wf`
+        * :py:func:`~qsiprep.workflows.anatomical.init_gifti_surface_wf`
+    """
+
+    workflow = Workflow(name=name)
+    workflow.__desc__ = """\
+Brain surfaces were reconstructed using `recon-all` [FreeSurfer {fs_ver},
+RRID:SCR_001847, @fs_reconall], and the brain mask estimated
+previously was refined with a custom variation of the method to reconcile
+ANTs-derived and FreeSurfer-derived segmentations of the cortical
+gray-matter of Mindboggle [RRID:SCR_002438, @mindboggle].
+""".format(fs_ver=fs.Info().looseversion() or '<ver>')
+
+    inputnode = pe.Node(
+        niu.IdentityInterface(
+            fields=['t1w', 't2w', 'flair', 'skullstripped_t1', 'corrected_t1', 'ants_segs',
+                    'subjects_dir', 'subject_id']), name='inputnode')
+    outputnode = pe.Node(
+        niu.IdentityInterface(
+            fields=['subjects_dir', 'subject_id', 't1_2_fsnative_forward_transform',
+                    't1_2_fsnative_reverse_transform', 'surfaces', 'out_brainmask',
+                    'out_aseg', 'out_aparc', 'out_report']),
+        name='outputnode')
+
+    recon_config = pe.Node(FSDetectInputs(hires_enabled=hires), name='recon_config')
+
+    autorecon1 = pe.Node(
+        fs.ReconAll(directive='autorecon1', flags='-noskullstrip', openmp=omp_nthreads),
+        name='autorecon1', n_procs=omp_nthreads, mem_gb=5)
+    autorecon1.interface._can_resume = False
+    autorecon1.interface._always_run = True
+
+    skull_strip_extern = pe.Node(FSInjectBrainExtracted(), name='skull_strip_extern')
+
+    fsnative_2_t1_xfm = pe.Node(RobustRegister(auto_sens=True, est_int_scale=True),
+                                name='fsnative_2_t1_xfm')
+    t1_2_fsnative_xfm = pe.Node(LTAConvert(out_lta=True, invert=True),
+                                name='t1_2_fsnative_xfm')
+
+    autorecon_resume_wf = init_autorecon_resume_wf(omp_nthreads=omp_nthreads)
+    gifti_surface_wf = init_gifti_surface_wf()
+
+    aseg_to_native_wf = init_segs_to_native_wf()
+    aparc_to_native_wf = init_segs_to_native_wf(segmentation='aparc_aseg')
+    refine = pe.Node(RefineBrainMask(), name='refine')
+
+    workflow.connect([
+        # Configuration
+        (inputnode, recon_config, [('t1w', 't1w_list'),
+                                   ('t2w', 't2w_list'),
+                                   ('flair', 'flair_list')]),
+        # Passing subjects_dir / subject_id enforces serial order
+        (inputnode, autorecon1, [('subjects_dir', 'subjects_dir'),
+                                 ('subject_id', 'subject_id')]),
+        (autorecon1, skull_strip_extern, [('subjects_dir', 'subjects_dir'),
+                                          ('subject_id', 'subject_id')]),
+        (skull_strip_extern, autorecon_resume_wf, [('subjects_dir', 'inputnode.subjects_dir'),
+                                                   ('subject_id', 'inputnode.subject_id')]),
+        (autorecon_resume_wf, gifti_surface_wf, [
+            ('outputnode.subjects_dir', 'inputnode.subjects_dir'),
+            ('outputnode.subject_id', 'inputnode.subject_id')]),
+        # Reconstruction phases
+        (inputnode, autorecon1, [('t1w', 'T1_files')]),
+        (recon_config, autorecon1, [('t2w', 'T2_file'),
+                                    ('flair', 'FLAIR_file'),
+                                    ('hires', 'hires'),
+                                    # First run only (recon-all saves expert options)
+                                    ('mris_inflate', 'mris_inflate')]),
+        (inputnode, skull_strip_extern, [('skullstripped_t1', 'in_brain')]),
+        (recon_config, autorecon_resume_wf, [('use_t2w', 'inputnode.use_T2'),
+                                             ('use_flair', 'inputnode.use_FLAIR')]),
+        # Construct transform from FreeSurfer conformed image to FMRIPREP
+        # reoriented image
+        (inputnode, fsnative_2_t1_xfm, [('t1w', 'target_file')]),
+        (autorecon1, fsnative_2_t1_xfm, [('T1', 'source_file')]),
+        (fsnative_2_t1_xfm, gifti_surface_wf, [
+            ('out_reg_file', 'inputnode.t1_2_fsnative_reverse_transform')]),
+        (fsnative_2_t1_xfm, t1_2_fsnative_xfm, [('out_reg_file', 'in_lta')]),
+        # Refine ANTs mask, deriving new mask from FS' aseg
+        (inputnode, refine, [('corrected_t1', 'in_anat'),
+                             ('ants_segs', 'in_ants')]),
+        (inputnode, aseg_to_native_wf, [('corrected_t1', 'inputnode.in_file')]),
+        (autorecon_resume_wf, aseg_to_native_wf, [
+            ('outputnode.subjects_dir', 'inputnode.subjects_dir'),
+            ('outputnode.subject_id', 'inputnode.subject_id')]),
+        (inputnode, aparc_to_native_wf, [('corrected_t1', 'inputnode.in_file')]),
+        (autorecon_resume_wf, aparc_to_native_wf, [
+            ('outputnode.subjects_dir', 'inputnode.subjects_dir'),
+            ('outputnode.subject_id', 'inputnode.subject_id')]),
+        (aseg_to_native_wf, refine, [('outputnode.out_file', 'in_aseg')]),
+
+        # Output
+        (autorecon_resume_wf, outputnode, [('outputnode.subjects_dir', 'subjects_dir'),
+                                           ('outputnode.subject_id', 'subject_id'),
+                                           ('outputnode.out_report', 'out_report')]),
+        (gifti_surface_wf, outputnode, [('outputnode.surfaces', 'surfaces')]),
+        (t1_2_fsnative_xfm, outputnode, [('out_lta', 't1_2_fsnative_forward_transform')]),
+        (fsnative_2_t1_xfm, outputnode, [('out_reg_file', 't1_2_fsnative_reverse_transform')]),
+        (refine, outputnode, [('out_file', 'out_brainmask')]),
+        (aseg_to_native_wf, outputnode, [('outputnode.out_file', 'out_aseg')]),
+        (aparc_to_native_wf, outputnode, [('outputnode.out_file', 'out_aparc')]),
+    ])
+
+    return workflow
+
+
+def init_autorecon_resume_wf(omp_nthreads, name='autorecon_resume_wf'):
+    r"""
+    This workflow resumes recon-all execution, assuming the `-autorecon1` stage
+    has been completed.
+    In order to utilize resources efficiently, this is broken down into five
+    sub-stages; after the first stage, the second and third stages may be run
+    simultaneously, and the fourth and fifth stages may be run simultaneously,
+    if resources permit::
+        $ recon-all -sd <output dir>/freesurfer -subjid sub-<subject_label> \
+            -autorecon2-volonly
+        $ recon-all -sd <output dir>/freesurfer -subjid sub-<subject_label> \
+            -autorecon-hemi lh \
+            -noparcstats -nocortparc2 -noparcstats2 -nocortparc3 \
+            -noparcstats3 -nopctsurfcon -nohyporelabel -noaparc2aseg \
+            -noapas2aseg -nosegstats -nowmparc -nobalabels
+        $ recon-all -sd <output dir>/freesurfer -subjid sub-<subject_label> \
+            -autorecon-hemi rh \
+            -noparcstats -nocortparc2 -noparcstats2 -nocortparc3 \
+            -noparcstats3 -nopctsurfcon -nohyporelabel -noaparc2aseg \
+            -noapas2aseg -nosegstats -nowmparc -nobalabels
+        $ recon-all -sd <output dir>/freesurfer -subjid sub-<subject_label> \
+            -autorecon3 -hemi lh -T2pial
+        $ recon-all -sd <output dir>/freesurfer -subjid sub-<subject_label> \
+            -autorecon3 -hemi rh -T2pial
+    The excluded steps in the second and third stages (``-no<option>``) are not
+    fully hemisphere independent, and are therefore postponed to the final two
+    stages.
+    .. workflow::
+        :graph2use: orig
+        :simple_form: yes
+        from qsiprep.workflows.anatomical import init_autorecon_resume_wf
+        wf = init_autorecon_resume_wf(omp_nthreads=1)
+    **Inputs**
+        subjects_dir
+            FreeSurfer SUBJECTS_DIR
+        subject_id
+            FreeSurfer subject ID
+        use_T2
+            Refine pial surface using T2w image
+        use_FLAIR
+            Refine pial surface using FLAIR image
+    **Outputs**
+        subjects_dir
+            FreeSurfer SUBJECTS_DIR
+        subject_id
+            FreeSurfer subject ID
+        out_report
+            Reportlet visualizing quality of surface alignment
+    """
+    workflow = Workflow(name=name)
+
+    inputnode = pe.Node(
+        niu.IdentityInterface(
+            fields=['subjects_dir', 'subject_id', 'use_T2', 'use_FLAIR']),
+        name='inputnode')
+
+    outputnode = pe.Node(
+        niu.IdentityInterface(
+            fields=['subjects_dir', 'subject_id', 'out_report']),
+        name='outputnode')
+
+    autorecon2_vol = pe.Node(
+        fs.ReconAll(directive='autorecon2-volonly', openmp=omp_nthreads),
+        n_procs=omp_nthreads, mem_gb=5, name='autorecon2_vol')
+    autorecon2_vol.interface._always_run = True
+
+    autorecon_surfs = pe.MapNode(
+        fs.ReconAll(
+            directive='autorecon-hemi',
+            flags=['-noparcstats', '-nocortparc2', '-noparcstats2',
+                   '-nocortparc3', '-noparcstats3', '-nopctsurfcon',
+                   '-nohyporelabel', '-noaparc2aseg', '-noapas2aseg',
+                   '-nosegstats', '-nowmparc', '-nobalabels'],
+            openmp=omp_nthreads),
+        iterfield='hemi', n_procs=omp_nthreads, mem_gb=5,
+        name='autorecon_surfs')
+    autorecon_surfs.inputs.hemi = ['lh', 'rh']
+    autorecon_surfs.interface._always_run = True
+
+    autorecon3 = pe.MapNode(
+        fs.ReconAll(directive='autorecon3', openmp=omp_nthreads),
+        iterfield='hemi', n_procs=omp_nthreads, mem_gb=5,
+        name='autorecon3')
+    autorecon3.inputs.hemi = ['lh', 'rh']
+    autorecon3.interface._always_run = True
+
+    # Only generate the report once; should be nothing to do
+    recon_report = pe.Node(
+        ReconAllRPT(directive='autorecon3', generate_report=True),
+        name='recon_report', mem_gb=5)
+    recon_report.interface._always_run = True
+
+    def _dedup(in_list):
+        vals = set(in_list)
+        if len(vals) > 1:
+            raise ValueError(
+                "Non-identical values can't be deduplicated:\n{!r}".format(in_list))
+        return vals.pop()
+
+    workflow.connect([
+        (inputnode, autorecon3, [('use_T2', 'use_T2'),
+                                 ('use_FLAIR', 'use_FLAIR')]),
+        (inputnode, autorecon2_vol, [('subjects_dir', 'subjects_dir'),
+                                     ('subject_id', 'subject_id')]),
+        (autorecon2_vol, autorecon_surfs, [('subjects_dir', 'subjects_dir'),
+                                           ('subject_id', 'subject_id')]),
+        (autorecon_surfs, autorecon3, [(('subjects_dir', _dedup), 'subjects_dir'),
+                                       (('subject_id', _dedup), 'subject_id')]),
+        (autorecon3, outputnode, [(('subjects_dir', _dedup), 'subjects_dir'),
+                                  (('subject_id', _dedup), 'subject_id')]),
+        (autorecon3, recon_report, [(('subjects_dir', _dedup), 'subjects_dir'),
+                                    (('subject_id', _dedup), 'subject_id')]),
+        (recon_report, outputnode, [('out_report', 'out_report')]),
+    ])
+
+    return workflow
+
+
+def init_gifti_surface_wf(name='gifti_surface_wf'):
+    r"""
+    This workflow prepares GIFTI surfaces from a FreeSurfer subjects directory
+    If midthickness (or graymid) surfaces do not exist, they are generated and
+    saved to the subject directory as ``lh/rh.midthickness``.
+    These, along with the gray/white matter boundary (``lh/rh.smoothwm``), pial
+    sufaces (``lh/rh.pial``) and inflated surfaces (``lh/rh.inflated``) are
+    converted to GIFTI files.
+    Additionally, the vertex coordinates are :py:class:`recentered
+    <qsiprep.interfaces.NormalizeSurf>` to align with native T1w space.
+    .. workflow::
+        :graph2use: orig
+        :simple_form: yes
+        from qsiprep.workflows.anatomical import init_gifti_surface_wf
+        wf = init_gifti_surface_wf()
+    **Inputs**
+        subjects_dir
+            FreeSurfer SUBJECTS_DIR
+        subject_id
+            FreeSurfer subject ID
+        t1_2_fsnative_reverse_transform
+            LTA formatted affine transform file (inverse)
+    **Outputs**
+        surfaces
+            GIFTI surfaces for gray/white matter boundary, pial surface,
+            midthickness (or graymid) surface, and inflated surfaces
+    """
+    workflow = Workflow(name=name)
+
+    inputnode = pe.Node(niu.IdentityInterface(['subjects_dir', 'subject_id',
+                                               't1_2_fsnative_reverse_transform']),
+                        name='inputnode')
+    outputnode = pe.Node(niu.IdentityInterface(['surfaces']), name='outputnode')
+
+    get_surfaces = pe.Node(nio.FreeSurferSource(), name='get_surfaces')
+
+    midthickness = pe.MapNode(
+        MakeMidthickness(thickness=True, distance=0.5, out_name='midthickness'),
+        iterfield='in_file',
+        name='midthickness')
+
+    save_midthickness = pe.Node(nio.DataSink(parameterization=False),
+                                name='save_midthickness')
+
+    surface_list = pe.Node(niu.Merge(4, ravel_inputs=True),
+                           name='surface_list', run_without_submitting=True)
+    fs_2_gii = pe.MapNode(fs.MRIsConvert(out_datatype='gii'),
+                          iterfield='in_file', name='fs_2_gii')
+    fix_surfs = pe.MapNode(NormalizeSurf(), iterfield='in_file', name='fix_surfs')
+
+    workflow.connect([
+        (inputnode, get_surfaces, [('subjects_dir', 'subjects_dir'),
+                                   ('subject_id', 'subject_id')]),
+        (inputnode, save_midthickness, [('subjects_dir', 'base_directory'),
+                                        ('subject_id', 'container')]),
+        # Generate midthickness surfaces and save to FreeSurfer derivatives
+        (get_surfaces, midthickness, [('smoothwm', 'in_file'),
+                                      ('graymid', 'graymid')]),
+        (midthickness, save_midthickness, [('out_file', 'surf.@graymid')]),
+        # Produce valid GIFTI surface files (dense mesh)
+        (get_surfaces, surface_list, [('smoothwm', 'in1'),
+                                      ('pial', 'in2'),
+                                      ('inflated', 'in3')]),
+        (save_midthickness, surface_list, [('out_file', 'in4')]),
+        (surface_list, fs_2_gii, [('out', 'in_file')]),
+        (fs_2_gii, fix_surfs, [('converted', 'in_file')]),
+        (inputnode, fix_surfs, [('t1_2_fsnative_reverse_transform', 'transform_file')]),
+        (fix_surfs, outputnode, [('out_file', 'surfaces')]),
+    ])
+    return workflow
+
+
+def init_segs_to_native_wf(name='segs_to_native', segmentation='aseg'):
+    """
+    Get a segmentation from FreeSurfer conformed space into native T1w space
+    .. workflow::
+        :graph2use: orig
+        :simple_form: yes
+        from qsiprep.workflows.anatomical import init_segs_to_native_wf
+        wf = init_segs_to_native_wf()
+    **Parameters**
+        segmentation
+            The name of a segmentation ('aseg' or 'aparc_aseg' or 'wmparc')
+    **Inputs**
+        in_file
+            Anatomical, merged T1w image after INU correction
+        subjects_dir
+            FreeSurfer SUBJECTS_DIR
+        subject_id
+            FreeSurfer subject ID
+    **Outputs**
+        out_file
+            The selected segmentation, after resampling in native space
+    """
+    workflow = Workflow(name='%s_%s' % (name, segmentation))
+    inputnode = pe.Node(niu.IdentityInterface([
+        'in_file', 'subjects_dir', 'subject_id']), name='inputnode')
+    outputnode = pe.Node(niu.IdentityInterface(['out_file']), name='outputnode')
+    # Extract the aseg and aparc+aseg outputs
+    fssource = pe.Node(nio.FreeSurferSource(), name='fs_datasource')
+    tonative = pe.Node(fs.Label2Vol(), name='tonative')
+    tonii = pe.Node(fs.MRIConvert(out_type='niigz', resample_type='nearest'), name='tonii')
+
+    if segmentation.startswith('aparc'):
+        if segmentation == 'aparc_aseg':
+            def _sel(x): return [parc for parc in x if 'aparc+' in parc][0]
+        elif segmentation == 'aparc_a2009s':
+            def _sel(x): return [parc for parc in x if 'a2009s+' in parc][0]
+        elif segmentation == 'aparc_dkt':
+            def _sel(x): return [parc for parc in x if 'DKTatlas+' in parc][0]
+        segmentation = (segmentation, _sel)
+
+    workflow.connect([
+        (inputnode, fssource, [
+            ('subjects_dir', 'subjects_dir'),
+            ('subject_id', 'subject_id')]),
+        (inputnode, tonii, [('in_file', 'reslice_like')]),
+        (fssource, tonative, [(segmentation, 'seg_file'),
+                              ('rawavg', 'template_file'),
+                              ('aseg', 'reg_header')]),
+        (tonative, tonii, [('vol_label_file', 'in_file')]),
+        (tonii, outputnode, [('out_file', 'out_file')]),
+    ])
+    return workflow
+
+
+def init_anat_reports_wf(reportlets_dir, output_spaces, force_spatial_normalization,
+                         template, freesurfer, name='anat_reports_wf'):
+    """
+    Set up a battery of datasinks to store reports in the right location
+    """
+    workflow = Workflow(name=name)
+
+    inputnode = pe.Node(
+        niu.IdentityInterface(
+            fields=['source_file', 't1_conform_report', 'seg_report',
+                    't1_2_mni_report', 'recon_report']),
+        name='inputnode')
+
+    ds_t1_conform_report = pe.Node(
+        DerivativesDataSink(base_directory=reportlets_dir, suffix='conform'),
+        name='ds_t1_conform_report', run_without_submitting=True)
+
+    ds_t1_2_mni_report = pe.Node(
+        DerivativesDataSink(base_directory=reportlets_dir, suffix='t1_2_mni'),
+        name='ds_t1_2_mni_report', run_without_submitting=True)
+
+    ds_t1_seg_mask_report = pe.Node(
+        DerivativesDataSink(base_directory=reportlets_dir, suffix='seg_brainmask'),
+        name='ds_t1_seg_mask_report', run_without_submitting=True)
+
+    ds_recon_report = pe.Node(
+        DerivativesDataSink(base_directory=reportlets_dir, suffix='reconall'),
+        name='ds_recon_report', run_without_submitting=True)
+
+    workflow.connect([
+        (inputnode, ds_t1_conform_report, [('source_file', 'source_file'),
+                                           ('t1_conform_report', 'in_file')]),
+        (inputnode, ds_t1_seg_mask_report, [('source_file', 'source_file'),
+                                            ('seg_report', 'in_file')]),
+    ])
+
+    if freesurfer:
+        workflow.connect([
+            (inputnode, ds_recon_report, [('source_file', 'source_file'),
+                                          ('recon_report', 'in_file')])
+        ])
+    if 'template' in output_spaces or force_spatial_normalization:
+        workflow.connect([
+            (inputnode, ds_t1_2_mni_report, [('source_file', 'source_file'),
+                                             ('t1_2_mni_report', 'in_file')])
+        ])
+
+    return workflow
+
+
+def init_anat_derivatives_wf(output_dir, output_spaces, template, freesurfer,
+                             force_spatial_normalization, name='anat_derivatives_wf'):
+    """
+    Set up a battery of datasinks to store derivatives in the right location
+    """
+    workflow = Workflow(name=name)
+
+    inputnode = pe.Node(
+        niu.IdentityInterface(
+            fields=['source_files', 't1_template_transforms',
+                    't1_preproc', 't1_mask', 't1_seg', 't1_tpms',
+                    't1_2_mni_forward_transform', 't1_2_mni_reverse_transform',
+                    't1_2_mni', 'mni_mask', 'mni_seg', 'mni_tpms',
+                    't1_2_fsnative_forward_transform', 'surfaces',
+                    't1_fs_aseg', 't1_fs_aparc']),
+        name='inputnode')
+
+    t1_name = pe.Node(niu.Function(function=fix_multi_T1w_source_name), name='t1_name')
+
+    ds_t1_preproc = pe.Node(
+        DerivativesDataSink(base_directory=output_dir, desc='preproc', keep_dtype=True),
+        name='ds_t1_preproc', run_without_submitting=True)
+
+    ds_t1_mask = pe.Node(
+        DerivativesDataSink(base_directory=output_dir, desc='brain', suffix='mask'),
+        name='ds_t1_mask', run_without_submitting=True)
+
+    ds_t1_seg = pe.Node(
+        DerivativesDataSink(base_directory=output_dir, suffix='dseg'),
+        name='ds_t1_seg', run_without_submitting=True)
+
+    ds_t1_tpms = pe.Node(
+        DerivativesDataSink(base_directory=output_dir,
+                            suffix='label-{extra_value}_probseg'),
+        name='ds_t1_tpms', run_without_submitting=True)
+    ds_t1_tpms.inputs.extra_values = ['CSF', 'GM', 'WM']
+
+    ds_t1_mni = pe.Node(
+        DerivativesDataSink(base_directory=output_dir,
+                            space=template, desc='preproc', keep_dtype=True),
+        name='ds_t1_mni', run_without_submitting=True)
+
+    ds_mni_mask = pe.Node(
+        DerivativesDataSink(base_directory=output_dir,
+                            space=template, desc='brain', suffix='mask'),
+        name='ds_mni_mask', run_without_submitting=True)
+
+    ds_mni_seg = pe.Node(
+        DerivativesDataSink(base_directory=output_dir,
+                            space=template, suffix='dseg'),
+        name='ds_mni_seg', run_without_submitting=True)
+
+    ds_mni_tpms = pe.Node(
+        DerivativesDataSink(base_directory=output_dir,
+                            space=template, suffix='label-{extra_value}_probseg'),
+        name='ds_mni_tpms', run_without_submitting=True)
+    ds_mni_tpms.inputs.extra_values = ['CSF', 'GM', 'WM']
+
+    # Transforms
+    suffix_fmt = 'from-{}_to-{}_mode-image_xfm'.format
+    ds_t1_mni_inv_warp = pe.Node(
+        DerivativesDataSink(base_directory=output_dir,
+                            suffix=suffix_fmt(template, 'T1w')),
+        name='ds_t1_mni_inv_warp', run_without_submitting=True)
+
+    ds_t1_template_transforms = pe.MapNode(
+        DerivativesDataSink(base_directory=output_dir, suffix=suffix_fmt('orig', 'T1w')),
+        iterfield=['source_file', 'in_file'],
+        name='ds_t1_template_transforms', run_without_submitting=True)
+
+    ds_t1_mni_warp = pe.Node(
+        DerivativesDataSink(
+            base_directory=output_dir,
+            suffix=suffix_fmt('T1w', template)),
+        name='ds_t1_mni_warp', run_without_submitting=True)
+
+    workflow.connect([
+        (inputnode, t1_name, [('source_files', 'in_files')]),
+        (inputnode, ds_t1_template_transforms, [('source_files', 'source_file'),
+                                                ('t1_template_transforms', 'in_file')]),
+        (inputnode, ds_t1_preproc, [('t1_preproc', 'in_file')]),
+        (inputnode, ds_t1_mask, [('t1_mask', 'in_file')]),
+        (inputnode, ds_t1_seg, [('t1_seg', 'in_file')]),
+        (inputnode, ds_t1_tpms, [('t1_tpms', 'in_file')]),
+        (t1_name, ds_t1_preproc, [('out', 'source_file')]),
+        (t1_name, ds_t1_mask, [('out', 'source_file')]),
+        (t1_name, ds_t1_seg, [('out', 'source_file')]),
+        (t1_name, ds_t1_tpms, [('out', 'source_file')]),
+    ])
+
+    if 'template' in output_spaces or force_spatial_normalization:
+        workflow.connect([
+            (inputnode, ds_t1_mni_warp, [('t1_2_mni_forward_transform', 'in_file')]),
+            (inputnode, ds_t1_mni_inv_warp, [('t1_2_mni_reverse_transform', 'in_file')]),
+            (inputnode, ds_t1_mni, [('t1_2_mni', 'in_file')]),
+            (inputnode, ds_mni_mask, [('mni_mask', 'in_file')]),
+            (inputnode, ds_mni_seg, [('mni_seg', 'in_file')]),
+            (inputnode, ds_mni_tpms, [('mni_tpms', 'in_file')]),
+            (t1_name, ds_t1_mni_warp, [('out', 'source_file')]),
+            (t1_name, ds_t1_mni_inv_warp, [('out', 'source_file')]),
+            (t1_name, ds_t1_mni, [('out', 'source_file')]),
+            (t1_name, ds_mni_mask, [('out', 'source_file')]),
+            (t1_name, ds_mni_seg, [('out', 'source_file')]),
+            (t1_name, ds_mni_tpms, [('out', 'source_file')]),
+        ])
+
+    return workflow
+
+
+def _seg2msks(in_file, newpath=None):
+    """Converts labels to masks"""
+    import nibabel as nb
+    import numpy as np
+    from nipype.utils.filemanip import fname_presuffix
+
+    nii = nb.load(in_file)
+    labels = nii.get_fdata()
+
+    out_files = []
+    for i in range(1, 4):
+        ldata = np.zeros_like(labels)
+        ldata[labels == i] = 1
+        out_files.append(fname_presuffix(
+            in_file, suffix='_label%03d' % i, newpath=newpath))
+        nii.__class__(ldata, nii.affine, nii.header).to_filename(out_files[-1])
+
+    return out_files
+
+
+    contrast_name = anatomical_contrast.lower()
+    if not infant_mode:
+        ref_img = pkgr('qsiprep',
+                       'data/mni_1mm_%s_lps.nii.gz' % contrast_name)
+        ref_img_brain = pkgr('qsiprep',
+                             'data/mni_1mm_%s_lps_brain.nii.gz' % contrast_name)
+    else:
+        ref_img = pkgr('qsiprep',
+                       'data/mni_1mm_%s_lps_infant.nii.gz' % contrast_name)
+        ref_img_brain = pkgr('qsiprep',
+                             'data/mni_1mm_%s_lps_brain_infant.nii.gz' % contrast_name)
+
+    return ref_img, ref_img_brain
